+++ Expression.h
@@ -9,8 +9,8 @@ using namespace std;
 class Expression {
 
 public:
-    virtual double calculate() = 0;
-    virtual ~Expression() {}
+    virtual float calculate() = 0;
+    virtual ~Expression(){}
 };
 
 #endif //PROJECT_EXPRESSION_H
